+++ qlab/problems/__init__.py
@@ -17,8 +17,9 @@ from qlab.problems import (  # noqa: F401  (import = registration)
     shor,
     simon,
     state_prep,
+    teleportation,
     vqe,
 )
 
 __all__ = ["state_prep", "maxcut", "bernstein_vazirani", "deutsch_jozsa", "simon", "grover", "qft", "qpe",
-           "shor", "vqe", "qml_classifier", "noise", "qec_repetition", "qec_surface", "chsh"]
+           "shor", "vqe", "qml_classifier", "noise", "qec_repetition", "qec_surface", "chsh", "teleportation"]